guard against empty name
tilemaker changed recently such that `Attribute('foo', '')` writes an
empty string. Before it would just ignore it.

I think the idea is that you should use `Has('foo')` first.

Without this change, we write a bunch of `name=''` attributes, which
screws up our rendering for things that just test for the existence of
the name attribute.

Maybe a better fix would be to monkey-patch Attribute in all of my
scripts?